fix for system-test
