+++ src/entity.ts
@@ -19,6 +19,7 @@ import * as extend from 'extend';
 import * as is from 'is';
 import {Query, QueryProto} from './query';
 import {PathType} from '.';
+import {Entities} from './request';
 
 // tslint:disable-next-line no-namespace
 export namespace entity {
@@ -653,6 +654,55 @@ export namespace entity {
       ent[entity.KEY_SYMBOL] = entity.keyFromKeyProto(result.entity.key!);
       return ent;
     });
+  }
+
+  /**
+   * Find the properties which value size is large than 1500 bytes,
+   * with excludeLargeProperties enabled, automatically exclude properties from indexing.
+   * This will allow storing string values larger than 1500 bytes
+   *
+   * @param entities Datastore key object(s).
+   * @param path namespace of provided entity properties
+   * @param properties properties which value size is large than 1500 bytes
+   */
+  export function findLargeProperties_(
+    entities: Entities,
+    path: string,
+    properties: string[] = []
+  ) {
+    const MAX_DATASTORE_VALUE_LENGTH = 1500;
+    if (Array.isArray(entities)) {
+      for (const entry of entities) {
+        if (entry.hasOwnProperty('name') && entry.hasOwnProperty('value')) {
+          if (
+            typeof entry.value === 'string' &&
+            Buffer.from(entry.value).length > MAX_DATASTORE_VALUE_LENGTH
+          ) {
+            entry.excludeFromIndexes = true;
+          } else {
+            continue;
+          }
+        }
+        findLargeProperties_(entry, path.concat('[]'), properties);
+      }
+    } else if (typeof entities === 'object') {
+      const keys = Object.keys(entities);
+      for (const key of keys) {
+        findLargeProperties_(
+          entities[key],
+          path.concat(`${path ? '.' : ''}${key}`),
+          properties
+        );
+      }
+    } else if (
+      typeof entities === 'string' &&
+      Buffer.from(entities).length > MAX_DATASTORE_VALUE_LENGTH
+    ) {
+      if (properties.indexOf(path) < 0) {
+        properties.push(path);
+      }
+    }
+    return properties;
   }
 
   /**
